dead애니메이션 / 연출1, lerp
애니메이션은 좀 어색해서 추후 고칠 수도 있습니다

+++ src/manager/TextureManager.cpp
@@ -61,10 +61,10 @@ void TextureManager::drawFrame(string id, int x, int y, int width, int height, i
 	float m_percent = (2 * percent) * 0.01f;
 
 	srcRect.w = width;
-	srcRect.h = m_percent * height;
+	srcRect.h = (int)(m_percent * height);
 
 	destRect.w = width;
-	destRect.h = m_percent * height;
+	destRect.h = (int)(m_percent * height);
 
 	srcRect.x = width * currentFrame;
 	srcRect.y = height * currentRow;
